+++ src/Player.hs
@@ -6,11 +6,17 @@ module Player
     , GunState(..)
     , Aim2D(..)
     , Vitality(..)
+    , Movement(..)
     , playerMaxHealth
     , Deaths
     , ReloadTime
+    , BoostTime
     , playerSide
     , minFiringInterval
+    , boostRechargeTime
+    , boostDuration
+    , fullBoostCycle
+    , boostSpeed
     , playerSize
     , createPlayer
     , staticPlayerImage
@@ -59,25 +65,35 @@ import           Data.Word                      ( Word8 )
 import           Data.Int                       ( Int16 )
 import           SDL.Raw.Types                  ( JoystickID )
 import           Data.Maybe
-import           Data.Tuple.Extra               ( (&&&) )
 import           Data.List                      ( foldl' )
 
 type AxisId = Word8
 type AxisPosition = Int16
+type ButtonId = Word8
+data AxisEvent = AxisEvent AxisId AxisPosition deriving Eq
+data ButtonEvent = ButtonEvent ButtonId JoyButtonState deriving Eq
+type InputEvent = Either AxisEvent ButtonEvent
 
-type Direction1D = Float
 -- Keep a 2D vector to cover the case when the aim stick is only moved in one
 -- axis. Keep the rotation to cover the case when the aim stick is moved to the
 -- default position (0, 0).
 data Aim2D = Aim2D Direction1D Direction1D Angle2D deriving (Show, Eq)
-type ReloadTime = DeltaTime
+type ReloadTime = Time
 data GunState = Firing | Idle deriving (Show, Eq)
 data Gun = Gun Aim2D ReloadTime GunState deriving (Show, Eq)
-data Color = Red | Blue deriving (Show, Eq)
+
 type Deaths = Int
 data Vitality = Vitality Deaths Health deriving (Show, Eq)
+
+data Color = Red | Blue deriving (Show, Eq)
 data PlayerId = PlayerId Color (Maybe JoystickID) deriving (Show, Eq)
-data Player = Player Circle Velocity2D Gun Vitality PlayerId deriving (Show, Eq)
+
+type BoostTime = Time
+-- Keep the direction in addition to the velocity to record left trigger
+-- movements while in boost mode. In boost mode the velocity is locked.
+data Movement = Movement Velocity2D Direction2D BoostTime deriving (Show, Eq)
+
+data Player = Player Circle Movement Gun Vitality PlayerId deriving (Show, Eq)
 
 playerMaxHealth :: Health
 playerMaxHealth = 1.0
@@ -91,12 +107,26 @@ playerRadius = playerSide / 2
 minFiringInterval :: ReloadTime
 minFiringInterval = 250
 
+boostDuration :: BoostTime
+boostDuration = 300
+
+boostRechargeTime :: BoostTime
+boostRechargeTime = 3000
+
+fullBoostCycle :: BoostTime
+fullBoostCycle = boostDuration + boostRechargeTime
+
+boostSpeed :: Float
+boostSpeed = 1.6
+
 axisPositionToVelocity :: Float
 axisPositionToVelocity = 0.000018
 
 triggerMinFireValue = 0
 minAxisPosition = 5000
+leftBumberButtonId = 4
 rightBumberButtonId = 5
+leftTriggerButtonId = 2
 rightTriggerButtonId = 5
 baseColor = PixelRGBA8 0xE6 0xE6 0xE6 255
 loadShadeColor = PixelRGBA8 0x16 0x0f 0x35 120
@@ -109,7 +139,7 @@ playerSize = V2 playerSide playerSide
 createPlayer :: Position2D -> Angle2D -> Color -> Maybe JoystickID -> Player
 createPlayer pos direction color joystickId = Player
     (Circle pos (playerSide / 2))
-    (V2 0 0)
+    (Movement (V2 0 0) (V2 0 0) 0)
     (Gun (Aim2D 0 0 direction) 0 Idle)
     (Vitality 0 playerMaxHealth)
     (PlayerId color joystickId)
@@ -118,13 +148,15 @@ staticPlayerImage :: (ImageId, VectorImage)
 staticPlayerImage = (baseImageId, toSolidCircleImage baseColor playerRadius)
 
 drawPlayer :: Player -> [(Rectangle Float, Either VectorImage ImageId)]
-drawPlayer (Player outline _ gun (Vitality _ health) (PlayerId colorId _)) =
-    let Gun (Aim2D _ _ direction) reloadTime _ = gun
-        center        = R.V2 playerRadius playerRadius
-        textureArea   = toTextureArea outline
-        diameter      = toDiameter playerRadius
-        healthRadians = (1 - health) * (2 * pi)
-        shadowLength  = reloadTimeToAimShadowLength (4 / 3) reloadTime
+drawPlayer (Player outline movement gun vitality (PlayerId colorId _)) =
+    let center                        = R.V2 playerRadius playerRadius
+        textureArea                   = toTextureArea outline
+        diameter                      = toDiameter playerRadius
+        Vitality _ health             = vitality
+        healthRadians                 = (1 - health) * (2 * pi)
+        Movement _ _ boostTime        = movement
+        shadowLength = boostTimeToAimShadowLength (4 / 3) boostTime
+        Gun (Aim2D _ _ direction) _ _ = gun
     in  [ (textureArea, Right baseImageId)
         , ( textureArea
           , Left
@@ -148,9 +180,12 @@ aimColor :: Color -> PixelRGBA8
 aimColor Red  = PixelRGBA8 0xD3 0x5F 0x5F 255
 aimColor Blue = PixelRGBA8 0x5F 0x5F 0xD3 255
 
-reloadTimeToAimShadowLength :: Float -> ReloadTime -> Float
-reloadTimeToAimShadowLength maxShadowWidth reloadTime =
-    maxShadowWidth * fromIntegral reloadTime / fromIntegral minFiringInterval
+boostTimeToAimShadowLength :: Float -> BoostTime -> Float
+boostTimeToAimShadowLength maxShadowWidth boostTime =
+    maxShadowWidth * if boostTime > boostRechargeTime
+        then fromIntegral (fullBoostCycle - boostTime)
+            / fromIntegral boostDuration
+        else fromIntegral boostTime / fromIntegral boostRechargeTime
 
 --               , - ~ ~ ~ - ,
 --           , '               ' ,
@@ -187,7 +222,7 @@ aimShadowShape len
     = []
     | len > 0 && len < 1 / 3
     = let
-          t                  = 2 / pi * acos (3 * len)
+          t                  = len * 3
           cutTopLeftCurve    = snd $ breakCubicBezierAt topLeftCurve (1 - t)
           cutBottomLeftCurve = fst $ breakCubicBezierAt bottomLeftCurve t
       in
@@ -242,24 +277,127 @@ createVelocity x y | isCloseToDefault x && isCloseToDefault y = V2 0 0
 
 updatePlayer :: [Event] -> DeltaTime -> Obstacles -> Player -> Player
 updatePlayer events dt obstacles player =
-    let Player shape velocity gun vitality (PlayerId color maybeJoystickId)
-            = player
-        Gun aim reloadTime state = gun
-        axisEvents               = case maybeJoystickId of
-            Just joystickId ->
-                map (joyAxisEventAxis &&& joyAxisEventValue)
-                    $ filter ((joystickId ==) . joyAxisEventWhich)
-                    $ mapMaybe toJoyAxis events
-            Nothing -> []
-        newVelocity   = foldl' updateVelocity velocity axisEvents
-        newAim        = foldl' updateAim aim axisEvents
-        newCircle     = moveCollidingCircle newVelocity dt obstacles shape
-        newReloadTime = max 0 $ reloadTime - dt
-    in  Player newCircle
-               newVelocity
-               (Gun newAim newReloadTime state)
-               vitality
-               (PlayerId color $ updateJoystick events =<< maybeJoystickId)
+    let
+        Player shape movement gun vitality playerId = player
+        PlayerId color maybeJoystickId = playerId
+        playerEvents                   = getPlayerEvents maybeJoystickId events
+        axisEvents                     = toAxisEvents playerEvents
+        newGun                         = updateGun axisEvents dt gun
+        Gun newAim _ _                 = newGun
+        buttonEvents                   = toButtonEvents playerEvents
+        (newMovement, velocities) =
+            updateMovement dt buttonEvents axisEvents newAim movement
+        newCircle = updateCircle velocities obstacles shape
+        newPlayerId =
+            (PlayerId color $ updateJoystick events =<< maybeJoystickId)
+    in
+        Player newCircle newMovement newGun vitality newPlayerId
+
+updateGun :: [AxisEvent] -> DeltaTime -> Gun -> Gun
+updateGun axisEvents dt (Gun aim reloadTime state) =
+    Gun (foldl' updateAim aim axisEvents) (countDown reloadTime dt) state
+
+updateMovement
+    :: DeltaTime
+    -> [ButtonEvent]
+    -> [AxisEvent]
+    -> Aim2D
+    -> Movement
+    -> (Movement, [(Velocity2D, Time)])
+updateMovement dt buttonEvents axisEvents aim movement =
+    let Aim2D _ _ aimAngle = aim
+        Movement velocity direction boostTime = movement
+        newBoostTime = updateBoostTime dt buttonEvents axisEvents boostTime
+        newDirection = foldl' updateVelocity direction axisEvents
+        willNotBeBoosting = newBoostTime <= boostRechargeTime
+        newVelocity
+            | newBoostTime > boostTime = if newDirection /= V2 0 0
+                then boostSpeed @* toUnitVector newDirection
+                else boostSpeed @* angleToVector aimAngle
+            | willNotBeBoosting = newDirection
+            | otherwise = velocity
+        wasBoosting = boostTime >= boostRechargeTime
+        velocities  = if wasBoosting && willNotBeBoosting
+            then
+                [ (velocity   , boostTime - boostRechargeTime)
+                , (newVelocity, boostRechargeTime - newBoostTime)
+                ]
+            else [(newVelocity, dt)]
+    in  (Movement newVelocity newDirection newBoostTime, velocities)
+
+countDown :: Time -> DeltaTime -> Time
+countDown current dt = max 0 $ current - dt
+
+updateBoostTime
+    :: DeltaTime -> [ButtonEvent] -> [AxisEvent] -> BoostTime -> BoostTime
+updateBoostTime _ buttonEvents axisEvents 0 =
+    if any isLeftBumperPressed buttonEvents
+            || any isLeftTriggerPressed axisEvents
+        then fullBoostCycle
+        else 0
+updateBoostTime dt _ _ current = countDown current dt
+
+getPlayerEvents :: Maybe JoystickID -> [Event] -> [Event]
+getPlayerEvents Nothing _ = []
+getPlayerEvents (Just joystickId) events =
+    filter (hasJoystickId joystickId) events
+
+hasJoystickId :: JoystickID -> Event -> Bool
+hasJoystickId joystickId = maybe False (joystickId ==) . getJoystickId
+
+getJoystickId :: Event -> Maybe JoystickID
+getJoystickId (Event _ payload) = case payload of
+    JoyAxisEvent (JoyAxisEventData joyId _ _) -> Just joyId
+    JoyButtonEvent (JoyButtonEventData joyId _ _) -> Just joyId
+    JoyDeviceEvent (JoyDeviceEventData _ joyId) -> Just joyId
+    _ -> Nothing
+
+toAxisEvents :: [Event] -> [AxisEvent]
+toAxisEvents = mapMaybe toAxisEvent
+
+toButtonEvents :: [Event] -> [ButtonEvent]
+toButtonEvents = mapMaybe toButtonEvent
+
+toAxisEvent :: Event -> Maybe AxisEvent
+toAxisEvent (Event _ (JoyAxisEvent joyAxisEventData)) =
+    let JoyAxisEventData _ axisId position = joyAxisEventData
+    in  Just $ AxisEvent axisId position
+toAxisEvent _ = Nothing
+
+toButtonEvent :: Event -> Maybe ButtonEvent
+toButtonEvent (Event _ (JoyButtonEvent joyButtonEventData)) =
+    let JoyButtonEventData _ buttonId state = joyButtonEventData
+    in  Just $ ButtonEvent buttonId state
+toButtonEvent _ = Nothing
+
+toInputEvent :: Event -> Maybe InputEvent
+toInputEvent event = case toAxisEvent event of
+    Just axisEvent -> Just $ Left axisEvent
+    _              -> case toButtonEvent event of
+        Just buttonEvent -> Just $ Right buttonEvent
+        _                -> Nothing
+
+toInputEvents :: [Event] -> [InputEvent]
+toInputEvents = mapMaybe toInputEvent
+
+isLeftBumperPressed :: ButtonEvent -> Bool
+isLeftBumperPressed = (==) $ ButtonEvent leftBumberButtonId JoyButtonPressed
+
+isRightBumperPressed :: ButtonEvent -> Bool
+isRightBumperPressed = (==) $ ButtonEvent rightBumberButtonId JoyButtonPressed
+
+isRightBumperReleased :: ButtonEvent -> Bool
+isRightBumperReleased =
+    (==) $ ButtonEvent rightBumberButtonId JoyButtonReleased
+
+isLeftTriggerPressed :: AxisEvent -> Bool
+isLeftTriggerPressed (AxisEvent axisId axisPositon) =
+    axisId == leftTriggerButtonId && axisPositon > triggerMinFireValue
+
+updateCircle :: [(Velocity2D, DeltaTime)] -> Obstacles -> Circle -> Circle
+updateCircle [] _ shape = shape
+updateCircle ((velocity, dt) : vs) obstacles shape =
+    updateCircle vs obstacles $ moveCollidingCircle velocity dt obstacles shape
 
 updateJoystick :: [Event] -> JoystickID -> Maybe JoystickID
 updateJoystick events joystickId = if hasJoystickRemovedEvent events joystickId
@@ -273,20 +411,16 @@ hasJoystickRemovedEvent events joystickId = any
     )
     events
 
-toJoyAxis :: Event -> Maybe JoyAxisEventData
-toJoyAxis (Event _ (JoyAxisEvent joyAxisEventData)) = Just joyAxisEventData
-toJoyAxis _ = Nothing
-
-updateAim :: Aim2D -> (AxisId, AxisPosition) -> Aim2D
-updateAim (Aim2D x y direction) (axisId, axisPos) =
+updateAim :: Aim2D -> AxisEvent -> Aim2D
+updateAim (Aim2D x y direction) (AxisEvent axisId axisPos) =
     let pos = fromIntegral axisPos
     in  case axisId of
             3 -> Aim2D pos y $ createAngle pos y direction
             4 -> Aim2D x pos $ createAngle x pos direction
             _ -> Aim2D x y direction
 
-updateVelocity :: Velocity2D -> (AxisId, AxisPosition) -> Velocity2D
-updateVelocity (V2 x y) (axisId, axisPos) =
+updateVelocity :: Velocity2D -> AxisEvent -> Velocity2D
+updateVelocity (V2 x y) (AxisEvent axisId axisPos) =
     let newV = axisPositionToVelocity * fromIntegral axisPos
     in  case axisId of
             0 -> createVelocity newV y
@@ -296,9 +430,10 @@ updateVelocity (V2 x y) (axisId, axisPos) =
 fireBullet :: [Event] -> BulletId -> Player -> (Player, Maybe Bullet)
 fireBullet events bulletId player =
     let Player (Circle position _) _ gun _ (PlayerId _ joystickId) = player
-        Gun   aim reloadTime state     = gun
-        Aim2D _   _          direction = aim
-        newState                       = getGunState state joystickId events
+        Gun aim reloadTime state = gun
+        Aim2D _ _ direction = aim
+        playerEvents = getPlayerEvents joystickId events
+        newState = getGunState state $ toInputEvents playerEvents
         (newReloadTime, maybeBullet) = if reloadTime == 0 && newState == Firing
             then
                 ( minFiringInterval
@@ -307,27 +442,21 @@ fireBullet events bulletId player =
             else (reloadTime, Nothing)
     in  (setGun (Gun aim newReloadTime newState) player, maybeBullet)
 
-getGunState :: GunState -> Maybe JoystickID -> [Event] -> GunState
-getGunState gunState (Just joystickId) events =
-    foldl' seq gunState $ mapMaybe (eventToGunState joystickId) events
-getGunState gunState Nothing _ = gunState
+getGunState :: GunState -> [InputEvent] -> GunState
+getGunState gunState events =
+    foldl' seq gunState $ mapMaybe eventToGunState events
 
-eventToGunState :: JoystickID -> Event -> Maybe GunState
-eventToGunState playerJoystickId (Event _ (JoyAxisEvent axisEventData)) =
-    let JoyAxisEventData joystickId buttonId amountPressed = axisEventData
-    in  if joystickId == playerJoystickId && buttonId == rightTriggerButtonId
-            then if amountPressed > triggerMinFireValue
-                then Just Firing
-                else Just Idle
-            else Nothing
-eventToGunState playerJoystickId (Event _ (JoyButtonEvent buttonEventData)) =
-    let JoyButtonEventData joystickId buttonId buttonState = buttonEventData
-    in  if joystickId == playerJoystickId && buttonId == rightBumberButtonId
-            then case buttonState of
-                JoyButtonPressed  -> Just Firing
-                JoyButtonReleased -> Just Idle
-            else Nothing
-eventToGunState _ _ = Nothing
+eventToGunState :: InputEvent -> Maybe GunState
+eventToGunState (Left (AxisEvent axisId amountPressed)) =
+    if axisId == rightTriggerButtonId
+        then if amountPressed > triggerMinFireValue
+            then Just Firing
+            else Just Idle
+        else Nothing
+eventToGunState (Right buttonEvent)
+    | isRightBumperPressed buttonEvent  = Just Firing
+    | isRightBumperReleased buttonEvent = Just Idle
+    | otherwise                         = Nothing
 
 setGun :: Gun -> Player -> Player
 setGun gun (Player shape velocity _ vitality playerId) =