+++ CHANGELOG.md
@@ -7,8 +7,15 @@ and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Added player boost capabilities triggered with left bumper button or left
+  trigger button.
+
+### Changed
+* The gun reload animation is now used for boost recharge instead.
+
 ### Fixed
-* Improved performance for when drawing players
+* Improved performance for drawing players
 
 ## [0.3.0] - 2020-02-25
 ### Added